server/requirements.txt
django===2.2.5
django-crontab===0.7.1
djangorestframework===3.9.4